Added Blur and Noise to Generate
Added Blur and Noise Generation to Generation.cpp.
Also improved the Generations itself. Now each folder has evenly distributed images correspondig to the Generation (For example the rotations 45,90,135,180,225,270,315 are evenly distributed in 03_rotate).
Also the workingFiles are shuffled a few times leading to completely randomly chosen images, yet still satisfying the above. The reason for this is the fact that only choosing the skip-th image lead to non-random behaviour and could lead to only choosing very few images due to repitition.
Changed ScaleGeneration to only Scaling by 6.0 for now (the images did not differ too much).